+++ app/src/main/java/com/example/practise2048champs/enums/GameModes.java
@@ -13,138 +13,138 @@ import lombok.Getter;
 */
 @Getter
 public enum GameModes {
-    SQUARE_4X4(4, 4, "4 X 4", "SQUARE", 2048, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
+    SQUARE_4X4(4, 4, "4 X 4", "SQUARE", 2048L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
     }}, "square_4x4.png"), // Total 16 cells.
 
-    SQUARE_5X5(5,5,"5 X 5","SQUARE",4096, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
+    SQUARE_5X5(5,5,"5 X 5","SQUARE",4096L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
     }}, "square_5x5.png"), // Total 25 cells.
 
-    SQUARE_6X6(6,6,"6 X 6","SQUARE",8192, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
+    SQUARE_6X6(6,6,"6 X 6","SQUARE",8192L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
     }}, "square_6x6.png"), // Total 36 cells.
 
-    RECTANGLE_4X5(4,5,"4 X 5","RECTANGLE",2048, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
+    RECTANGLE_4X5(4,5,"4 X 5","RECTANGLE",2048L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
     }}, "rectangle_4x5.png"), // Total 20 cells.
 
-    RECTANGLE_4X6(4,6,"4 X 6","RECTANGLE",4096, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
+    RECTANGLE_4X6(4,6,"4 X 6","RECTANGLE",4096L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
     }}, "rectangle_4x6.png"), // Total 24 cells.
 
-    RECTANGLE_5X6(5,6,"5 X 6","RECTANGLE",8192, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
+    RECTANGLE_5X6(5,6,"5 X 6","RECTANGLE",8192L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
     }}, "rectangle_5x6.png"), // Total 30 cells.
 
-    BLOCK_MIDDLE_SQUARE_5X5(5, 5, "5 X 5", "BLOCK MIDDLE SQ.", 2048, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(-1); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
+    BLOCK_MIDDLE_SQUARE_5X5(5, 5, "5 X 5", "BLOCK MIDDLE SQ.", 2048L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(-1L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
     }}, "block_middle_square_5x5.png"), // Total 25 cells.
 
-    BLOCK_MIDDLE_SQUARE_6X6(6,6,"? X ?","BLOCK MIDDLE SQ.",4096, false, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(-1); add(-1); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(-1); add(-1); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
+    BLOCK_MIDDLE_SQUARE_6X6(6,6,"? X ?","BLOCK MIDDLE SQ.",4096L, false, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(-1L); add(-1L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(-1L); add(-1L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
     }}, "arriving_game_mode.png"), // Total 36 cells.
 
-    BLOCK_MIDDLE_RECTANGLE_4X6(4, 6, "4 X 6", "BLOCK MIDDLE REC.", 2048, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(-1); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(-1); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
+    BLOCK_MIDDLE_RECTANGLE_4X6(4, 6, "4 X 6", "BLOCK MIDDLE REC.", 2048L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(-1L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(-1L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
     }}, "block_middle_rectangle_4x6.png"), // Total 24 cells.
 
-    BLOCK_MIDDLE_RECTANGLE_5X6(5,6,"? X ?","BLOCK MIDDLE REC.",4096, false, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(-1); add(-1); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(-1); add(-1); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
+    BLOCK_MIDDLE_RECTANGLE_5X6(5,6,"? X ?","BLOCK MIDDLE REC.",4096L, false, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(-1L); add(-1L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(-1L); add(-1L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
     }}, "arriving_game_mode.png"), // Total 30 cells.
 
-    BLOCK_2_CORNERS_SQUARE_5X5(5, 5, "5 X 5", "BLOCK 2 CORNERS SQ.", 2048, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(-1); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(-1);}});
+    BLOCK_2_CORNERS_SQUARE_5X5(5, 5, "5 X 5", "BLOCK 2 CORNERS SQ.", 2048L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(-1L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(-1L);}});
     }}, "block_2_corners_square_5x5.png"), // Total 25 cells.
 
-    BLOCK_2_CORNERS_SQUARE_6X6(6, 6, "? X ?", "BLOCK 2 CORNERS SQ.", 4096, false, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(-1); add(-1); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(-1); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0); add(-1);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(-1); add(-1);}});
+    BLOCK_2_CORNERS_SQUARE_6X6(6, 6, "? X ?", "BLOCK 2 CORNERS SQ.", 4096L, false, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(-1L); add(-1L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(-1L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L); add(-1L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(-1L); add(-1L);}});
     }}, "arriving_game_mode.png"), // Total 36 cells.
 
-    BLOCK_2_CORNERS_RECTANGLE_4X6(4, 6, "4 X 6", "BLOCK 2 CORNERS REC.", 2048, true, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(-1); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(-1);}});
+    BLOCK_2_CORNERS_RECTANGLE_4X6(4, 6, "4 X 6", "BLOCK 2 CORNERS REC.", 2048L, true, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(-1L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(-1L);}});
     }}, "block_2_corners_rectangle_4x6.png"), // Total 24 cells.
 
-    BLOCK_2_CORNERS_RECTANGLE_5X6(5,6,"? X ?","BLOCK 2 CORNERS REC.",4096, false, new ArrayList<>() {{
-        add(new ArrayList<>() {{add(-1); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(-1); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(0);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(-1);}});
-        add(new ArrayList<>() {{add(0); add(0); add(0); add(0); add(-1);}});
+    BLOCK_2_CORNERS_RECTANGLE_5X6(5,6,"? X ?","BLOCK 2 CORNERS REC.",4096L, false, new ArrayList<>() {{
+        add(new ArrayList<>() {{add(-1L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(-1L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(0L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(-1L);}});
+        add(new ArrayList<>() {{add(0L); add(0L); add(0L); add(0L); add(-1L);}});
     }}, "arriving_game_mode.png"); // Total 30 cells.
 
     private final int columns;
     private final int rows;
     private final String dimensions;
     private final String mode;
-    private final int goal;
+    private final long goal;
     private final boolean canAccess;
-    private final ArrayList<ArrayList<Integer>> blockCells; // '-1' is for a block cell
+    private final ArrayList<ArrayList<Long>> blockCells; // '-1' is for a block cell
     private final String gamePreviewAssetFileName;
     private final GameLayoutProperties gameLayoutProperties;
 
-    GameModes(int columns, int rows, String dimensions, String mode, int goal, boolean canAccess,
-              ArrayList<ArrayList<Integer>> blockCells, String gamePreviewAssetFileName) {
+    GameModes(int columns, int rows, String dimensions, String mode, long goal, boolean canAccess,
+              ArrayList<ArrayList<Long>> blockCells, String gamePreviewAssetFileName) {
         this.columns = columns;
         this.rows = rows;
         this.dimensions = dimensions;
